apps/tg_news_bot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-news-bot"
version = "0.1.0"
description = "Telegram bot for news ingestion, moderation, and publishing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.4.1,<4.0",
  "aiohttp>=3.9.0,<4.0",
  "sqlalchemy>=2.0.25,<3.0",
  "asyncpg>=0.29.0,<1.0",
  "alembic>=1.13.1,<2.0",
  "httpx>=0.27.0,<1.0",
  "feedparser>=6.0.11,<7.0",
  "trafilatura>=1.9.0,<2.0",
  "readability-lxml>=0.8.1,<1.0",
  "lxml>=5.2.1,<6.0",
  "Pillow>=10.2.0,<11.0",
  "pydantic-settings>=2.2.1,<3.0",
  "structlog>=24.1.0,<25.0",
  "sentry-sdk>=2.0.0,<3.0",
  "apscheduler>=3.10.4,<4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1,<9.0",
  "pytest-asyncio>=0.23.6,<1.0"
]

[tool.setuptools.packages.find]
where = ["src"]
